Clean up the watcher code.
