+++ plugins/commands/watch.py
@@ -19,128 +19,31 @@
 # SOFTWARE.
 
 import atexit
+import contextlib
 import http.server
 import logging
 import os
 import signal
 import subprocess
 import sys
-import threading
-import time
 import urllib
 import webbrowser
 
 import watchdog.events
 import watchdog.observers
 
-import incontext
 import cli
+import incontext
 import paths
 import utils
 
-
-class CallbackEventHandler(watchdog.events.FileSystemEventHandler):
-    """Logs all the events captured."""
-
-    def __init__(self, callback):
-        self._callback = callback
-
-    def on_moved(self, event):
-        super(CallbackEventHandler, self).on_moved(event)
-        self._callback()
-
-    def on_created(self, event):
-        self._callback()
-
-    def on_deleted(self, event):
-        self._callback()
-
-    def on_modified(self, event):
-        self._callback()
-
-
-def watch_directory(paths, callback):
-    observer = watchdog.observers.Observer()
-    for path in paths:
-        observer.schedule(CallbackEventHandler(callback=callback), path, recursive=True)
-    observer.start()
-    return observer
-
-
-class Builder(threading.Thread):
-
-    def __init__(self, incontext, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.incontext = incontext
-        self.scheduled = False
-        self.lock = threading.Lock()
-        self.stopped = False
-
-    def schedule(self):
-        logging.info("Scheduling build...")
-        with self.lock:
-            self.scheduled = True
-
-    def stop(self):
-        logging.info("Stopping builder...")
-        with self.lock:
-            self.stopped = True
-
-    def run(self):
-        while True:
-            time.sleep(1)
-            scheduled = False
-            with self.lock:
-                if self.stopped:
-                    return
-                scheduled = self.scheduled
-                self.scheduled = False
-            if scheduled:
-                try:
-                    self.incontext.commands["build"].run()
-                    logging.info("Done.")
-                except Exception as e:
-                    logging.error("Failed: %s", e)
-
-
-def docker(command):
-    prefix = []
-    if sys.platform == "linux":
-        prefix = ["sudo"]
-    return subprocess.run(prefix + ["docker"] + command)
-
-
-class WatchingBuilder(object):
-
-    def __init__(self, incontext):
-        self.incontext = incontext
-
-    def start(self):
-        self.builder = Builder(self.incontext)
-        self.builder.start()
-        logging.info("Watching directory...")
-        self.observer = watch_directory([self.incontext.configuration.site.paths.content,
-                                         self.incontext.configuration.site.paths.templates],
-                                        self.builder.schedule)
-        logging.info("Performing initial build...")
-        self.builder.schedule()
-
-    def stop(self):
-        self.builder.stop()
-        self.observer.stop()
-        self.observer.join()
-        self.builder.join()
+import commands.build
 
 
 @incontext.command("watch", help="watch for changes and automatically build the website")
 def command_watch(incontext, options):
-    builder = WatchingBuilder(incontext)
-    builder.start()
-    try:
-        while True:
-            time.sleep(0.2)
-    except KeyboardInterrupt:
-        builder.stop()
+    with commands.build.WatchingBuilder(incontext):
+        utils.wait_for_keyboard_interrupt()
 
 
 @incontext.command("serve", help="run a local web server for development", arguments=[
@@ -152,18 +55,12 @@ def command_watch(incontext, options):
                  help="watch for changes and rebuild automatically")
 ])
 def command_serve(incontext, options):
-
-    builder = None
-    if options.watch:
-        builder = WatchingBuilder(incontext)
-        builder.start()
-
-    with utils.Chdir(incontext.configuration.site.destination.files_directory):
+    builder_context = commands.build.WatchingBuilder(incontext) if options.watch else contextlib.nullcontext()
+    with builder_context, utils.Chdir(incontext.configuration.site.destination.files_directory):
         httpd = http.server.HTTPServer(('', options.port),
                                        http.server.SimpleHTTPRequestHandler)
         logging.info("Listening on %s...", options.port)
         try:
             httpd.serve_forever()
         except KeyboardInterrupt:
-            if builder is not None:
-                builder.stop()
+            return
